added a trait for the commands and moved SnackFS to first level
